@@ -1,6 +1,6 @@
 {
   "name": "mapbox-gl-draw-scale-rotate",
-  "version": "1.1.0",
+  "version": "1.2.0",
   "description": "A custom mode for MapboxGL Draw to scale and rotate features",
   "main": "dist/index.js",
   "module": "dist/index.js",
@@ -28,29 +28,31 @@
     "dist"
   ],
   "devDependencies": {
-    "@mapbox/mapbox-gl-draw": "^1.4.1",
-    "@types/mapbox__mapbox-gl-draw": "^1.3.3",
-    "@typescript-eslint/eslint-plugin": "^5.58.0",
-    "@typescript-eslint/parser": "^5.58.0",
-    "esbuild": "^0.17.16",
+    "@eslint/eslintrc": "^3.3.1",
+    "@eslint/js": "^9.30.1",
+    "@stylistic/eslint-plugin": "^5.1.0",
+    "@types/mapbox__mapbox-gl-draw": "^1.4.9",
+    "@typescript-eslint/eslint-plugin": "^8.36.0",
+    "@typescript-eslint/parser": "^8.36.0",
+    "esbuild": "^0.25.6",
     "esbuild-plugin-copy": "^2.1.1",
-    "esbuild-plugin-d.ts": "^1.1.0",
-    "eslint": "^8.38.0",
-    "rimraf": "^5.0.0",
-    "typescript": "^5.0.4"
+    "esbuild-plugin-d.ts": "^1.3.1",
+    "eslint": "^9.30.1",
+    "rimraf": "^6.0.1",
+    "typescript": "^5.8.3",
+    "typescript-eslint": "^8.36.0"
   },
   "peerDependencies": {
-    "@mapbox/mapbox-gl-draw": ">=1.4.0"
+    "@mapbox/mapbox-gl-draw": ">=1.5.0"
   },
   "dependencies": {
-    "@turf/bearing": "^6.5.0",
-    "@turf/center": "^6.5.0",
-    "@turf/destination": "^6.5.0",
-    "@turf/distance": "^6.5.0",
-    "@turf/helpers": "^6.5.0",
-    "@turf/midpoint": "^6.5.0",
-    "@turf/transform-rotate": "^6.5.0",
-    "@turf/transform-scale": "^6.5.0",
-    "@turf/turf": "^6.5.0"
+    "@turf/bearing": ">=6.5.0",
+    "@turf/center": ">=6.5.0",
+    "@turf/destination": ">=6.5.0",
+    "@turf/distance": ">=6.5.0",
+    "@turf/helpers": ">=6.5.0",
+    "@turf/midpoint": ">=6.5.0",
+    "@turf/transform-rotate": ">=6.5.0",
+    "@turf/transform-scale": ">=6.5.0"
   }
 }
