@@ -1,9 +1,8 @@
-import { lib } from '@mapbox/mapbox-gl-draw';
+import MapboxGLDraw from '@mapbox/mapbox-gl-draw';
 
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
 export const ScaleRotateStyle: Array<{ [key: string]: any }> = [
-  ...lib.theme
-    .map(style => {
+  ...MapboxGLDraw.lib.theme
+    .map((style: any) => {
       switch (style.id) {
         case 'gl-draw-polygon-fill-inactive': return { ...style, filter: [...style.filter, ['!=', 'user_type', 'overlay']] };
         case 'gl-draw-polygon-fill-active': return { ...style, filter: [...style.filter, ['!=', 'user_type', 'overlay']] };
@@ -27,13 +26,13 @@ export const ScaleRotateStyle: Array<{ [key: string]: any }> = [
       ['==', 'active', 'false'],
       ['==', '$type', 'Polygon'],
       ['==', 'user_type', 'overlay'],
-      ['!=', 'mode', 'static']
+      ['!=', 'mode', 'static'],
     ],
     paint: {
       'fill-color': '#3bb2d0',
       'fill-outline-color': '#3bb2d0',
-      'fill-opacity': 0.01
-    }
+      'fill-opacity': 0.01,
+    },
   },
   {
     id: 'gl-draw-overlay-polygon-fill-active',
@@ -42,13 +41,13 @@ export const ScaleRotateStyle: Array<{ [key: string]: any }> = [
       'all',
       ['==', 'active', 'true'],
       ['==', '$type', 'Polygon'],
-      ['==', 'user_type', 'overlay']
+      ['==', 'user_type', 'overlay'],
     ],
     paint: {
       'fill-color': '#fbb03b',
       'fill-outline-color': '#fbb03b',
-      'fill-opacity': 0.01
-    }
+      'fill-opacity': 0.01,
+    },
   },
   {
     id: 'gl-draw-polygon-and-line-vertex-scale-icon',
@@ -58,22 +57,22 @@ export const ScaleRotateStyle: Array<{ [key: string]: any }> = [
       ['==', 'meta', 'vertex'],
       ['==', '$type', 'Point'],
       ['!=', 'mode', 'static'],
-      ['has', 'heading']
+      ['has', 'heading'],
     ],
     layout: {
       'icon-image': 'scale',
       'icon-allow-overlap': true,
       'icon-ignore-placement': true,
       'icon-rotation-alignment': 'map',
-      'icon-rotate': ['get', 'heading']
+      'icon-rotate': ['get', 'heading'],
     },
     paint: {
       'icon-opacity': 1.0,
       'icon-opacity-transition': {
         delay: 0,
-        duration: 0
-      }
-    }
+        duration: 0,
+      },
+    },
   },
   {
     id: 'gl-draw-line-rotate-point',
@@ -83,17 +82,17 @@ export const ScaleRotateStyle: Array<{ [key: string]: any }> = [
       ['==', 'meta', 'midpoint'],
       ['==', 'icon', 'rotate'],
       ['==', '$type', 'LineString'],
-      ['!=', 'mode', 'static']
+      ['!=', 'mode', 'static'],
     ],
     layout: {
       'line-cap': 'round',
-      'line-join': 'round'
+      'line-join': 'round',
     },
     paint: {
       'line-color': '#fbb03b',
       'line-dasharray': [0.2, 2],
-      'line-width': 2
-    }
+      'line-width': 2,
+    },
   },
   {
     id: 'gl-draw-polygon-rotate-point-stroke',
@@ -103,12 +102,12 @@ export const ScaleRotateStyle: Array<{ [key: string]: any }> = [
       ['==', 'meta', 'midpoint'],
       ['==', 'icon', 'rotate'],
       ['==', '$type', 'Point'],
-      ['!=', 'mode', 'static']
+      ['!=', 'mode', 'static'],
     ],
     paint: {
       'circle-radius': 4,
-      'circle-color': '#fff'
-    }
+      'circle-color': '#fff',
+    },
   },
   {
     id: 'gl-draw-polygon-rotate-point',
@@ -118,12 +117,12 @@ export const ScaleRotateStyle: Array<{ [key: string]: any }> = [
       ['==', 'meta', 'midpoint'],
       ['==', 'icon', 'rotate'],
       ['==', '$type', 'Point'],
-      ['!=', 'mode', 'static']
+      ['!=', 'mode', 'static'],
     ],
     paint: {
       'circle-radius': 2,
-      'circle-color': '#fbb03b'
-    }
+      'circle-color': '#fbb03b',
+    },
   },
   {
     id: 'gl-draw-polygon-rotate-point-icon',
@@ -133,21 +132,21 @@ export const ScaleRotateStyle: Array<{ [key: string]: any }> = [
       ['==', 'meta', 'midpoint'],
       ['==', 'icon', 'rotate'],
       ['==', '$type', 'Point'],
-      ['!=', 'mode', 'static']
+      ['!=', 'mode', 'static'],
     ],
     layout: {
       'icon-image': 'rotate',
       'icon-allow-overlap': true,
       'icon-ignore-placement': true,
       'icon-rotation-alignment': 'map',
-      'icon-rotate': ['get', 'heading']
+      'icon-rotate': ['get', 'heading'],
     },
     paint: {
       'icon-opacity': 1.0,
       'icon-opacity-transition': {
         delay: 0,
-        duration: 0
-      }
-    }
-  }
+        duration: 0,
+      },
+    },
+  },
 ];
